Move `Dashboard.tsx` from `components` to `pages`
...for it is a page, not a component.
